backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
playwright==1.49.1
